clippy lint fix

+++ src/lib.rs
@@ -571,7 +571,7 @@ impl Parser {
                     }
                 };
                 if let Some(float_value) = float_value {
-                    let scaled_value = float_value as f64 * signal_def.factor + signal_def.offset;
+                    let scaled_value = float_value * signal_def.factor + signal_def.offset;
                     return Some(DecodedSignal {
                         name: signal_def.name.clone(),
                         value: DecodedSignalValue::Numeric(scaled_value),
